Redone nvim config
