+++ lua/config/autocmd.lua
@@ -113,6 +113,11 @@ vim.api.nvim_create_autocmd("LspAttach", {
 		-- local navic = require("nvim-navic")
 		-- if client and client.server_capabilities.documentSymbolProvider then
 		-- 	navic.attach(client, event.buf)
-		-- end
+		vim.lsp.handlers["textDocument/hover"] = vim.lsp.with(vim.lsp.handlers.hover, {
+			border = "rounded",
+		}) -- end
+		vim.lsp.handlers["textDocument/signatureHelp"] = vim.lsp.with(vim.lsp.handlers.signature_help, {
+			border = "rounded",
+		})
 	end,
 })
