+++ lua/plugins/autocomplete.lua
@@ -6,15 +6,6 @@ return { -- Autocompletion
 		{
 			"L3MON4D3/LuaSnip",
 			build = "make install_jsregexp",
-			-- build = (function()
-			-- 	-- Build Step is needed for regex support in snippets.
-			-- 	-- This step is not supported in many windows environments.
-			-- 	-- Remove the below condition to re-enable on windows.
-			-- 	-- if vim.fn.has("win32") == 1 or vim.fn.executable("make") == 0 then
-			-- 	-- 	return
-			-- 	-- end
-			-- 	return "make install_jsregexp"
-			-- end)(),
 			dependencies = {
 				-- `friendly-snippets` contains a variety of premade snippets.
 				--    See the README about individual language/framework/plugin snippets:
@@ -97,27 +88,6 @@ return { -- Autocompletion
 						fallback()
 					end
 				end, { "i", "s" }),
-				-- Think of <c-l> as moving to the right of your snippet expansion.
-				--  So if you have a snippet that's like:
-				--  function $name($args)
-				--    $body
-				--  end
-				--
-				-- <c-l> will move you to the right of each of the expansion locations.
-				-- <c-h> is similar, except moving you backwards.
-				-- ["<C-l>"] = cmp.mapping(function()
-				-- 	if luasnip.expand_or_locally_jumpable() then
-				-- 		luasnip.expand_or_jump()
-				-- 	end
-				-- end, { "i", "s" }),
-				-- ["<C-h>"] = cmp.mapping(function()
-				-- 	if luasnip.locally_jumpable(-1) then
-				-- 		luasnip.jump(-1)
-				-- 	end
-				-- end, { "i", "s" }),
-
-				-- For more advanced Luasnip keymaps (e.g. selecting choice nodes, expansion) see:
-				--    https://github.com/L3MON4D3/LuaSnip?tab=readme-ov-file#keymaps
 			}),
 			sources = {
 				{ name = "nvim_lsp" },
