+++ lua/plugins/base.lua
@@ -1,130 +1,7 @@
 return {
 	"tpope/vim-sleuth",
-
-	-- "gc" to comment visual regions/lines
-	{ "numToStr/Comment.nvim", opts = {} },
-
-	{
-		"numToStr/Comment.nvim",
-		lazy = false,
-		config = function()
-			---@diagnostic disable-next-line: missing-fields
-			require("Comment").setup({
-				pre_hook = require("ts_context_commentstring.integrations.comment_nvim").create_pre_hook(),
-			})
-		end,
-	},
-	{ "JoosepAlviste/nvim-ts-context-commentstring" },
-
-	{
-		"lewis6991/gitsigns.nvim",
-		opts = {
-			signs = {
-				add = { text = "+" },
-				change = { text = "~" },
-				delete = { text = "_" },
-				topdelete = { text = "‾" },
-				changedelete = { text = "~" },
-			},
-			on_attach = function(buffnr)
-				local gs = package.loaded.gitsigns
-				local function nmap(mode, l, r, opts)
-					opts = opts or {}
-					opts.buffer = buffnr
-					vim.keymap.set(mode, l, r, opts)
-				end
-
-				nmap("n", "<leader>gb", function()
-					gs.blame_line({ full = false })
-				end, { desc = "git [b]lame line" })
-				nmap("n", "<leader>gd", gs.diffthis, { desc = "git [d]iff against index" })
-				nmap("n", "<leader>gD", function()
-					gs.diffthis("~")
-				end, { desc = "git [D]iff against last commit" })
-			end,
-		},
-	},
-	-- Highlight todo, notes, etc in comments
-	{
-		"folke/todo-comments.nvim",
-		event = "VimEnter",
-		dependencies = { "nvim-lua/plenary.nvim" },
-		opts = { signs = false },
-	},
-	{ -- Collection of various small independent plugins/modules
-		"echasnovski/mini.nvim",
-		config = function()
-			-- Better Around/Inside textobjects
-			--
-			-- Examples:
-			--  - va)  - [V]isually select [A]round [)]paren
-			--  - yinq - [Y]ank [I]nside [N]ext [']quote
-			--  - ci'  - [C]hange [I]nside [']quote
-			require("mini.ai").setup({ n_lines = 500 })
-
-			-- Add/delete/replace surroundings (brackets, quotes, etc.)
-			--
-			-- - saiw) - [S]urround [A]dd [I]nner [W]ord [)]Paren
-			-- - sd'   - [S]urround [D]elete [']quotes
-			-- - sr)'  - [S]urround [R]eplace [)] [']
-			require("mini.surround").setup()
-
-			-- Simple and easy statusline.
-			--  You could remove this setup call if you don't like it,
-			--  and try some other statusline plugin
-			local statusline = require("mini.statusline")
-			-- set use_icons to true if you have a Nerd Font
-			statusline.setup({ use_icons = vim.g.have_nerd_font })
-
-			-- You can configure sections in the statusline by overriding their
-			-- default behavior. For example, here we set the section for
-			-- cursor location to LINE:COLUMN
-			---@diagnostic disable-next-line: duplicate-set-field
-			statusline.section_location = function()
-				return "%2l:%-2v"
-			end
-		end,
-	},
 	"tpope/vim-fugitive",
 	"tpope/vim-rhubarb",
-	{
-		"m4xshen/autoclose.nvim",
-		config = function()
-			require("autoclose").setup({
-				options = {
-					disable_when_touch = true,
-				},
-			})
-		end,
-	},
-	{
-		"zbirenbaum/copilot.lua",
-		cmd = "Copilot",
-		build = ":Copilot auth",
-		enabled = true,
-		lazy = false,
-		autostart = true,
-		opts = {
-			enabled = true,
-			suggestion = {
-				enabled = true,
-				auto_trigger = false,
-				debounce = 75,
-				keymap = {
-					accept = "<C-x>",
-					dismiss = "<esc>",
-					next = "<C-z>",
-					prev = "<C-S-Z>",
-				},
-			},
-			panel = { enabled = false },
-			filetypes = {
-				markdown = true,
-				help = true,
-			},
-		},
-	},
-
 	{
 		"windwp/nvim-ts-autotag",
 		config = function()
@@ -157,54 +34,6 @@ return {
 		opts = {},
 	},
 	{
-		"folke/noice.nvim",
-		event = "VeryLazy",
-		opts = {
-			-- add any options here
-		},
-		dependencies = {
-			-- if you lazy-load any plugin below, make sure to add proper `module="..."` entries
-			"MunifTanjim/nui.nvim",
-		},
-		config = function()
-			require("noice").setup({
-				lsp = {
-					-- override markdown rendering so that **cmp** and other plugins use **Treesitter**
-					override = {
-						["vim.lsp.util.convert_input_to_markdown_lines"] = true,
-						["vim.lsp.util.stylize_markdown"] = true,
-						["cmp.entry.get_documentation"] = false, -- requires hrsh7th/nvim-cmp
-					},
-					signature = {
-						enabled = true,
-						auto_open = {
-							enabled = true,
-							trigger = true,
-							luasnip = true,
-							throttle = 50,
-						},
-					},
-				},
-				-- you can enable a preset for easier configuration
-				presets = {
-					bottom_search = true, -- use a classic bottom cmdline for search
-					command_palette = false, -- position the cmdline and popupmenu together
-					long_message_to_split = true, -- long messages will be sent to a split
-					inc_rename = false, -- enables an input dialog for inc-rename.nvim
-					lsp_doc_border = true, -- add a border to hover docs and signature help
-				},
-				cmdline = {
-					view = "cmdline",
-					format = {
-						cmdline = false,
-						lua = false,
-						help = false,
-					},
-				},
-			})
-		end,
-	},
-	{
 		"folke/trouble.nvim",
 		dependencies = { "nvim-tree/nvim-web-devicons" },
 		opts = {
@@ -224,11 +53,10 @@ return {
 		config = function()
 			require("aerial").setup({
 				filter_kind = false,
+				close_on_select = true,
+				autojump = true,
 			})
 		end,
-	},
-	{
-		"yioneko/nvim-vtsls",
 	},
 	{
 		"MeanderingProgrammer/render-markdown.nvim",
@@ -236,5 +64,9 @@ return {
 		-- dependencies = { "nvim-treesitter/nvim-treesitter", "echasnovski/mini.nvim" }, -- if you use the mini.nvim suite
 		-- dependencies = { 'nvim-treesitter/nvim-treesitter', 'echasnovski/mini.icons' }, -- if you use standalone mini plugins
 		dependencies = { "nvim-treesitter/nvim-treesitter", "nvim-tree/nvim-web-devicons" }, -- if you prefer nvim-web-devicons
+	},
+	{
+		"lambdalisue/suda.vim",
+		lazy = false,
 	},
 }
