+++ lua/plugins/lint.lua
@@ -8,9 +8,6 @@ return {
 			lint.linters_by_ft = {
 				markdown = { "markdownlint-cli2" },
 				-- javascript = { "eslint_d" }, handled by eslint-lsp
-				-- typescript = { "eslint_d" },
-				-- javascriptreact = { "eslint_d" },
-				-- typescriptreact = { "eslint_d" },
 			}
 
 			-- To allow other plugins to add linters to require('lint').linters_by_ft,
