+++ lua/plugins/whichkey.lua
@@ -4,16 +4,6 @@ return {
 	config = function()
 		local wk = require("which-key")
 		wk.setup()
-		-- require("which-key").register({
-		-- 	["<leader>c"] = { name = "[c]ode", _ = "which_key_ignore" },
-		-- 	["<leader>r"] = { name = "[r]ename", _ = "which_key_ignore" },
-		-- 	["<leader>s"] = { name = "[s]earch", _ = "which_key_ignore" },
-		-- 	["<leader>p"] = { name = "[p]roject", _ = "which_key_ignore" },
-		-- 	["<leader>d"] = { name = "[d]iagnostics", _ = "which_key_ignore" },
-		-- 	["<leader>g"] = { name = "[g]it", _ = "which_key_ignore" },
-		-- 	["<leader>h"] = { name = "[h]arpoon", _ = "which_key_ignore" },
-		-- 	["<leader>t"] = { name = "[t]odo", _ = "which_key_ignore" },
-		-- })
 		require("which-key").add({
 			{ "<leader>c", group = "[c]ode" },
 			{ "<leader>c_", hidden = true },
